deps: Unpin Solana dependencies
#### Problem

It's very useful to run downstream tests with newer versions of the
Solana crates to catch unwanted breakage. Mollusk pins some of its
Solana crate dependencies to 2.1.0.

The token repo uses mollusk for its testing framework, but because of
the pinning in mollusk, it's impossible to run downstream tests on token
with newer versions of the SDK.

#### Summary of changes

This might be totally off-base, and I might be missing something
important, but it'll be way more flexible and easier if mollusk unpins
its dependencies.